+++ docs/changelog.rst
@@ -15,6 +15,7 @@ New features
 ------------
 
 * Attribute ``clim`` in :class:`aspecd.plotting.SurfaceProperties`
+* Parameter ``threshold`` for determining the levels of a contour plot in class :class:`aspecd.plotting.SinglePlotter2D`
 
 
 Fixes